spider_platform-master/node/requirements.txt
aiohttp==3.6.2
amqp==2.5.2
aniso8601==8.0.0
APScheduler==3.6.3
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
billiard==3.6.3.0
celery==4.4.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-JSON==0.3.4
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpath==0.82
kombu==4.6.8
lxml==4.5.0
MarkupSafe==1.1.1
more-itertools==8.2.0
multidict==4.7.3
mysql-connector-python==8.0.20
packaging==20.3
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pycparser==2.20
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
pytest==5.4.1
pytest-base-url==1.4.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
retrying==1.3.3
selenium==3.141.0
six==1.14.0
SQLAlchemy==1.3.16
tzlocal==2.0.0
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.9
Werkzeug==1.0.1
WTForms==2.2.1
yarl==1.4.2
zipp==2.2.0
